CXO-7813 Create JUnit tests for every kind of identifier + Added unit test which tests every LoadProfileIdentifier created inside mdc.device.data and protocols (upl-impl) + Added unit test which uses reflection to search for all implementations of LoadProfileIdentifier and which tests if proper resolving code is present inside the LoadProfileServiceImpl

+++ upl/com.energyict.mdc.upl.impl/protocolcommon/src/main/java/com/energyict/protocolimplv2/identifiers/LoadProfileIdentifierById.java
@@ -23,7 +23,7 @@ import java.util.Set;
 @XmlRootElement
 public class LoadProfileIdentifierById implements LoadProfileIdentifier {
 
-    private final int loadProfileId;
+    private final long loadProfileId;
     private final ObisCode profileObisCode;
     private final DeviceIdentifier deviceIdentifier;
 
@@ -35,7 +35,7 @@ public class LoadProfileIdentifierById implements LoadProfileIdentifier {
         this.deviceIdentifier = null;
     }
 
-    public LoadProfileIdentifierById(int loadProfileId, ObisCode profileObisCode, DeviceIdentifier deviceIdentifier) {
+    public LoadProfileIdentifierById(long loadProfileId, ObisCode profileObisCode, DeviceIdentifier deviceIdentifier) {
         super();
         this.loadProfileId = loadProfileId;
         this.profileObisCode = profileObisCode;
@@ -54,7 +54,7 @@ public class LoadProfileIdentifierById implements LoadProfileIdentifier {
     }
 
     @XmlAttribute
-    public int getLoadProfileId() {
+    public long getLoadProfileId() {
         return loadProfileId;
     }
